% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_landing_event}
\alias{get_landing_event}
\title{Retrieve Event Information from the Landing input}
\usage{
get_landing_event(
  inputId,
  event_type,
  values,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{<\code{character}> The identifier for the landing id object in the Shiny application.}

\item{event_type}{<\code{character}>The type of event to retrieve information for.}

\item{values}{<\verb{character vector}> A vector of values to be retrieved from the
specified event.}

\item{session}{<\code{session}> The Shiny session object. Defaults to the current reactive
domain.}
}
\value{
The specified values for the event if it exists and is of the
specified event type, NULL otherwise.
}
\description{
This function retrieves the specified event information.
The type of event is specified with \code{event_type} and the values to be
retrieved are defined in \code{values}.
}
