% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing_events.R
\name{get_landing_news}
\alias{get_landing_news}
\title{Retrieve News Link Click Information from Landing input}
\usage{
get_landing_news(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{A \code{character} string that identifies the landing id object in
the Shiny application.}

\item{session}{The Shiny \code{session} object. By default, it uses the current
reactive domain.}
}
\value{
Returns the ID of the news card clicked if it exists. If it does not
exist, it returns NULL.
}
\description{
This function retrieves page link click information using \code{get_landing_event}.
}
