% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_drop_events.R
\name{get_theme_drop_click}
\alias{get_theme_drop_click}
\title{Retrieve Page Link Click Information from Theme dropdown input}
\usage{
get_theme_drop_click(inputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{A \code{character} string that identifies the theme dropdown id object in
the Shiny application.}

\item{session}{The Shiny \code{session} object. By default, it uses the current
reactive domain.}
}
\value{
Returns the ID of the page clicked if it exists. If it does not
exist, it returns NULL.
}
\description{
This function retrieves page link click information using \code{get_theme_drop_event}.
}
