% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{landing_input}
\alias{landing_input}
\title{Generate a Landing Input for a Curbcut app}
\usage{
landing_input(
  inputId,
  pages,
  c_city_svg,
  news_cards,
  discover_cards,
  team_cards,
  contributors,
  collabs,
  translation_df
)
}
\arguments{
\item{inputId}{<\code{character}> The ID of the input element.}

\item{pages}{<\code{data.frame}>
A data frame containing information about the pages. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each page. (\code{alp})
\item \code{theme}: Theme of the page (\code{Health})
\item \code{nav_title}: Navigation title of the page. (\verb{Active living potential})
}}

\item{c_city_svg}{<\code{character}>
A character string representing the top-left corner SVG image of the city.}

\item{news_cards}{<\code{data.frame}>
A data frame containing information about news cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each news card. (\code{census})
\item \code{icon}: Icon of the news card. (\code{demographics})
\item \code{title_en}: Title of the news card in English. (\verb{2021 Census})
\item \code{title_fr}: Title of the news card in French. (\verb{Recensement de 2021})
\item \code{text_en}: Text content of the news card in English.
\item \code{text_fr}: Text content of the news card in French.
}}

\item{discover_cards}{<\code{data.frame}>
A data frame containing information about discover cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each discover card. (\code{metro})
\item \code{img}: Image file path for the discover card. (\code{www/discover/metro.png})
\item \code{theme}: Theme of the discover card. (\code{Transport})
\item \code{en}: Text content of the discover card in English. (\verb{The Evolution of the Montreal metro})
\item \code{fr}: Text content of the discover card in French. (\verb{L'évolution du métro de Montréal})
}}

\item{team_cards}{<\code{data.frame}>
A data frame containing information about team cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each team card. (\code{maxbdb})
\item \code{img}: Image file path or URL for the team card. (\code{www/maxime_belanger_de_blois.png})
\item \code{name}: Name of the team member. (\verb{Maxime Bélanger De Blois})
\item \code{role_en}: Role of the team member in English. (\verb{Head of Technology and Data})
\item \code{role_fr}: Role of the team member in French. (\verb{Responsable technologie et données})
\item \code{bio_en}: Biography of the team member in English.
\item \code{bio_fr}: Biography of the team member in French.
\item \code{theme}: Theme of the team card. (\code{Health})
}}

\item{contributors}{<\verb{character vector}>
A character vector containing the names of contributors.}

\item{collabs}{<\code{data.frame}>
A data frame containing information about collaborators. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each collaborator. (\code{MSSI})
\item \code{img}: Image file path or URL for the collaborator. (\code{www/mcgill-logo.png})
\item \code{name}: Name of the collaborator. (\verb{McGill Sustainability Systems Initiative})
}}

\item{translation_df}{<\code{data.frame}>
A data frame containing translation pairs. It should have the following columns:
\itemize{
\item \code{en}: Text content in English. (\code{Housing}, ...)
\item \code{fr}: Text content in French. (\code{Logement}, ...)
}}
}
\value{
A ReactShiny input object.
}
\description{
The \code{landing_input} function generates a 'ReactShiny' input object for a
Curbcut app. It encodes image data into base64 format and prepares several
types of card data to be used in the landing page of the app.
}
