% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_box.R
\name{title_box_input}
\alias{title_box_input}
\title{Create a Themed Title Box Shiny Input}
\usage{
title_box_input(
  inputId,
  theme,
  title_text_title,
  title_text_main,
  title_text_extra,
  lang = "en",
  minWidth = "400px",
  maxWidth = "40vw",
  show = "false"
)
}
\arguments{
\item{inputId}{<\code{character}> An identifier string for the new Shiny input.
This ID makes it possible to refer to the input in server logic.}

\item{theme}{<\code{character}> A character string defining the theme for the title box.}

\item{title_text_title}{<\code{character}> A character string defining the main title text.}

\item{title_text_main}{<\code{character}> A character string defining the subtitle text.}

\item{title_text_extra}{<\code{character}> A character string defining any additional text.}

\item{minWidth}{<\code{character}> A character string specifying the minimum width of the title
box, defaults to '300px'.}

\item{maxWidth}{<\code{character}> A character string specifying the maximum width of the title
box, defaults to '400px'.}

\item{show}{<\code{character}> A character string indicating whether to show the title box by
default, 'true' or 'false'. Defaults to 'false'.}

\item{lan}{<\code{character}>g A character string defining the language for the title box,
defaults to English ('en').}
}
\value{
Returns a Shiny input element that can be added to a Curbcut page.
}
\description{
This function creates a React-based Shiny Input for title boxes.
}
