% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{update_landing}
\alias{update_landing}
\title{Update the landing page input}
\usage{
update_landing(session, inputId, configuration = NULL)
}
\arguments{
\item{session}{<\code{shiny.session}> The Shiny session within which to update the input value.}

\item{inputId}{<\code{character}> The ID of the input element to be updated.}

\item{configuration}{<\verb{names list}> The new configuration for the input
element.}
}
\value{
The function doesn't return a value but modifies the input element
state in the active Shiny session.
}
\description{
The \code{update_landing} function updates a 'ReactShiny' input object for Curbcut.
It allows changing the value of the input element with the given 'inputId'.
Optionally, it also updates the configuration of the input element if provided.
}
\seealso{
\code{\link{landing_input}} for creating the landing input.
}
