% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_drop.R
\name{update_theme_drop}
\alias{update_theme_drop}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
update_theme_drop(session, inputId, value, configuration = NULL)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
