% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_drop.R
\name{update_theme_drop_lang}
\alias{update_theme_drop_lang}
\title{Update the theme pages}
\usage{
update_theme_drop_lang(session, inputId, lang)
}
\arguments{
\item{session}{<\code{shiny.session}> The Shiny session within which to update
the theme pages.}

\item{inputId}{<\code{character}> The ID of the input element to be updated.}

\item{lang}{<\code{character}> Default language, 'en' or 'fr'.}
}
\value{
The function doesn't return a value but modifies the theme pages
in the active theme dropdown input in the active Shiny session.
}
\description{
The \code{update_theme_pages} function updates the pages in a 'ReactShiny' input
object for Curbcut, specifically in the theme dropdown. This can be useful
when there's a need to modify the page content in the active theme dropdown
during a session (for language change, for example.
}
