% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{landing_input}
\alias{landing_input}
\title{Generate a Landing Input for a Curbcut app}
\usage{
landing_input(
  inputId,
  h1_first_line,
  h1_second_line,
  pages,
  c_city_svg,
  news_cards = NULL,
  discover_cards,
  team_cards,
  contributors,
  collabs,
  translation_df,
  lang = "en",
  placeholder_video_src,
  video_src,
  show_cities = TRUE
)
}
\arguments{
\item{inputId}{<\code{character}> The ID of the input element.}

\item{h1_first_line}{<\code{character}> A list, with language, of what will be on
the first line of the H1 header. EXPLORE THE \code{X} REGION. ex.
\code{list(en = "MONTREAL", fr = "RÉGION DE")}}

\item{h1_second_line}{<\code{character}> A list, with language, of what will be on
the second line of the H1 header. EXPLORE THE MONTREAL \code{X}, ex.
\code{list(en = "REGION", fr = "MONTRÉAL")}}

\item{pages}{<\code{data.frame}>
A data frame containing information about the pages. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each page. (\code{alp})
\item \code{theme}: Theme of the page (\code{Health})
\item \code{nav_title}: Navigation title of the page. (\verb{Active living potential})
}}

\item{c_city_svg}{<\code{character}>
A character string representing the path to the top-left corner SVG image
of the city. (\code{c-montreal})}

\item{news_cards}{<\code{data.frame}>
A data frame containing information about news cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each news card. (\code{census})
\item \code{icon}: Icon of the news card. (\code{demographics})
\item \code{title_en}: Title of the news card in English. (\verb{2021 Census})
\item \code{title_fr}: Title of the news card in French. (\verb{Recensement de 2021})
\item \code{text_en}: Text content of the news card in English.
\item \code{text_fr}: Text content of the news card in French.
}}

\item{discover_cards}{<\code{data.frame}>
A data frame containing information about discover cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each discover card. (\code{metro})
\item \code{img}: base64 image for the discover card. (\code{www/discover/metro.png})
\item \code{theme}: Theme of the discover card. (\code{Transport})
\item \code{en}: Text content of the discover card in English. (\verb{The Evolution of the Montreal metro})
\item \code{fr}: Text content of the discover card in French. (\verb{L'évolution du métro de Montréal})
}}

\item{team_cards}{<\code{data.frame}>
A data frame containing information about team cards. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each team card. (\code{maxbdb})
\item \code{img}: base64 image for the team card. (\code{www/maxime_belanger_de_blois.png})
\item \code{name}: Name of the team member. (\verb{Maxime Bélanger De Blois})
\item \code{role_en}: Role of the team member in English. (\verb{Head of Technology and Data})
\item \code{role_fr}: Role of the team member in French. (\verb{Responsable technologie et données})
\item \code{theme}: Theme of the team card. (\code{Health})
}}

\item{contributors}{<\verb{character vector}>
A character vector containing the names of contributors.}

\item{collabs}{<\code{data.frame}>
A data frame containing information about collaborators. It should have the following columns:
\itemize{
\item \code{id}: Unique identifier for each collaborator. (\code{MSSI})
\item \code{img}: base64 image for the collaborator. (\code{www/mcgill-logo.png})
\item \code{name}: Name of the collaborator. (\verb{McGill Sustainability Systems Initiative})
}}

\item{translation_df}{<\code{data.frame}>
A data frame containing translation pairs. It should have the following columns:
\itemize{
\item \code{en}: Text content in English. (\code{Housing}, ...)
\item \code{fr}: Text content in French. (\code{Logement}, ...)
}}

\item{lang}{<\code{character}> Default language, 'en' or 'fr'.}

\item{placeholder_video_src}{<\code{character}> External link to a publicly available
mp4 video. The video will be used as the placeholder, until the user click on
to watch the intro video. e.g. \verb{https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_placeholder.mp4}}

\item{video_src}{<\verb{named list}> Every video must be named by the language it uses.
The videos are as external link to a publicly available mp4 video. This is the
full length official video the user will see when they click on to watch the
intro video. e.g. \code{list(en = "https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_en.mp4")}}

\item{show_cities}{<\code{logical}> Should we be showing the list of Curbcut Cities
in the footer? Defaults to TRUE.}
}
\value{
A ReactShiny input object.
}
\description{
The \code{landing_input} function generates a 'ReactShiny' input object for a
Curbcut app. It encodes image data into base64 format and prepares several
types of card data to be used in the landing page of the app.
}
